Add workaround for clicks going to the wrong heading depending on the theme.
Each post heading has a jumptarget element attached to it. On condensed themes
this element will be larger than the parent and can cover multiple headings.
This caused a click on one heading to be sent to another heading's click
handler via jumptarget.

Add a workaround by finding which header that was actually clicked by using
the event coordinates.